change components to correct file
